fix: Mails are sent to the organizer twice
The `ics` package doesn't support setting the `SCHEDULE-AGENT`
parameter, there is a PR (https://github.com/adamgibbons/ics/pull/248),
but it is not merged.

This is a workaround that relies on the fact that the `ics` package does
not properly escape the `name` field of the organizer. In a perfect
world, they would merge the PR and create a new release, but this is
unfortunately not the world we live in.

The `SCHEDULE-AGENT` tells the CalDav server that the invitation has
been sent by the client (which is cal in this case), preventing the
CalDav server to not send invitations itself.

refs #9485