Please review! next-auth-options.ts refactor + passkeys support
This refactors the very important `authorize` function. I'm not sure if we want to keep it like this, but the authorize function itself does a lot of checks, only some relevant to `credentials`.

Instead of duplicating it, I added a second parameter, `unsafe`, intended to only be used from server code like in the PasskeyProvider.

When `unsafe` is used, all the checks dependent on `credentials` are not being run anymore. However, I am not sure how safe this is, or if a larger refactor is needed that properly separates checks dependent on `credentials` from all other checks.