fix: Do not fallback to existing db value if input is null for settings.
Fixes #21412.
For `finalSeatsPerTimeSlot` and `finalRecurringEvent` we were falling back
to the DB value if `seatsPerTimeSlot` or `recurringEvent` was falsy. Instead,
we should only fallback to the DB value if those values are undefined.
Those values being undefined means that they have not changed, while those
values being null means they are being reset to not have any value in them.
This was causing an error when a user was trying to turn off `Recurring event`
and turn on `Offer seats` in the same event.